test: split bridge progress tests by owner
Why:
- Continue the test architecture migration by moving package-owned bridge, speech, and progress coverage out of the app-host test bundle.
- Reduce simulator/app-host cost without weakening Android bridge and progress behavior contracts.

What Changed:
- Split AndBibleTests+BridgeAndProgress into BibleView bridge dispatch tests, BibleCore speech/progress store tests, and BibleUI reader progress bridge tests.
- Added package-local BibleUI and BibleCore helpers for bridge emissions and in-memory settings stores.
- Removed the migrated app-host test file from the Xcode project and updated the migration plan.

Validation:
- plutil -lint AndBible.xcodeproj/project.pbxproj.
- git diff --cached --check.
- xcodebuild test for BridgeDispatchTests in BibleViewTests.
- xcodebuild test for SpeakServiceTests and ProgressStoreTests in BibleCoreTests.
- xcodebuild test for ReaderProgressBridgeTests in BibleUITests.
- Full BibleViewTests, BibleCoreTests, and BibleUITests simulator package suites passed after isolated reruns for process-restart artifacts.
- xcodebuild build-for-testing for the AndBible app scheme.
- swift package describe --type json.
- python3 scripts/check_repo_standards.py docblocks --all-files.
- python3 -m unittest discover -s scripts -p test_*.py.
- gitnexus detect_changes on staged worktree changes: low risk, no affected execution flows.

Impact:
- Test-only migration; production app behavior is unchanged.
- App-host test surface shrinks while preserving the migrated bridge/progress contracts in package-owned lanes.
